Changing Directory of Saving the Cast file
Earlier, the cast file was saved in the directory where the `recordterm` command has been run. Now, it is changed. The temporary cast file will be saved to the home directory of the current user. From here, SVG/GIF is created, and the created file is moved and the residual cast file is removed.